Move tests to `AccountSettingsTest`
`AccountManagerSettingsStoreTest` only contained tests that belong into `AccountSettingsTest`.
